requirements.txt
PyQt6>=6.5.0
pika>=1.3.0
pyserial>=3.5
requests>=2.31.0

